Make VEVENT objects visible which don't have a DTEND field
When importing VEVENT objects into ownCloud which don't have a DTEND field
but a DURATION field, the end of the event is not stored in the database and
thus the event is never shown to the user.

This patch calculates DTEND from DTSTART+DURATION in case DTEND isn't
available. This function (getDTEndFromVEvent) was already implemented, but not
used in this stage.

Applying this patch doesn't automatically reveal all events.