added more and optional colors
Based on the discussion in issue #460 more colors are defined.
Especially darkend variants are included now.
Basis is a ColorWheel with 12 segments + Gray. Starting from that pure
colors, darker and less saturated variants are generated in steps of
10%.

As not everybody may like that much colors, most of them are commented
out. But as they are provided with the source, it is easy to enable them
just by removing the comment.